requirements.txt
fastapi
uvicorn[standard]
pyannote.audio==3.1.1